% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGEobj.utils.R
\docType{package}
\name{DGEobj.utils-package}
\alias{DGEobj.utils-package}
\title{DGEobj.utils Package Overview}
\description{
TBD
}
\section{More Information}{

\code{browseVignettes(package = 'DGEobj.utils')}
}

