% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEdgeRNorm.R
\name{runEdgeRNorm}
\alias{runEdgeRNorm}
\title{Run edgeR normalization on DGEobj}
\usage{
runEdgeRNorm(dgeObj, normMethod = "TMM", plotFile = TRUE, plotLabels)
}
\arguments{
\item{dgeObj}{A DGEobj containing counts, design data, and gene annotation.}

\item{normMethod}{One of "TMM", "RLE", "upperquartile", or "none". (Default = "TMM")}

\item{plotFile}{Enable a bar plot of the norm.factors produced. (Default = TRUE)}

\item{plotLabels}{Sample text labels for the plot. Length must equal the number of
samples. (Default: sample number will be displayed)}
}
\value{
A DGEobj with a normalized DGEList added.
}
\description{
Returns a DGEobj containing DGEList object representing the result of
edgeR normalizationn (calcNormFactors).
}
\examples{
   myDGEobj <- readRDS(system.file("exampleObj.RDS", package = "DGEobj"))
   myDGEobj <- DGEobj::resetDGEobj(myDGEobj)

   # Default TMM normalization
   myDGEobj <- runEdgeRNorm(myDGEobj)

   # Set some options
   myDGEobj <- runEdgeRNorm(myDGEobj,
                            normMethod = "upperquartile",
                            plotFile = FALSE,
                            plotLabels = myDGEobj$design$Sample.name)

}
