% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyIntensity.R
\name{tidyIntensity}
\alias{tidyIntensity}
\title{Create a tidy intensity dataframe}
\usage{
tidyIntensity(intensityObj, rowIDcolname, keyColname, valueColname, group)
}
\arguments{
\item{intensityObj}{A dataframe or matrix with row names and col names. (Required)}

\item{rowIDcolname}{Column name for the rowID (Usually a geneID column).
If it doesn't exist, it will be created from row names. (Required)}

\item{keyColname}{Defines the column name that will hold the sample names
captured from the col names of the input dataframe.  (Required)}

\item{valueColname}{Defines the column name for the value column (intensity values). (Required)}

\item{group}{A grouping variable to define which columns go together (vector
with length = ncol(intensityObj) of the input dataframe).  This is typically the name
of a column from the design table. (Required)}
}
\value{
A tidy dataframe of intensity data.
}
\description{
This function takes a dataframe or matrix of intensity values with geneID row names
and sample name column names, e.g. from log2cpm.  It uses tidyr::gather to produce a
"tidy" dataframe.
}
\details{
The rowIDcolname defines the name the resulting gene id column should have and will create the
column from the row names, which are typically ensemble gene IDs. If a gene symbol column is
already present in the intensity dataframe, it can be specified using rowIDcolname and will
be used instead of the rownames.
}
\examples{
\dontrun{
  myTidyIntensity <- tidyIntensity(myLog2CPM,
                                   rowIDcolname = "GeneID",
                                   keyColname = "Sample"
                                   valueColname = "Log2CPM",
                                   group = dgeObj$design$ReplicateGroup)
}

}
